% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ams_id_ontology.R
\name{read_ams_id_technology_shapes}
\alias{read_ams_id_technology_shapes}
\title{Load AMS-I.D technology applicability shape}
\usage{
read_ams_id_technology_shapes()
}
\value{
A \code{sh_shape_graph} as returned by \code{\link[shapeR:read_shacl]{shapeR::read_shacl()}}.
}
\description{
Load AMS-I.D technology applicability shape
}
